Working preprocessing of data
Including the removal of invalid datapoints and temperature-correction
of load-data. Allows for easy pipelining of additional preprocessing
steps if wanted. Corrected the fact that old code did not take leap
years into account, as well as correcting for temperature deviations not
only in the winter-months (as described in Tønne PhD). Code may be
optimized, but runs currently quickly for dataset of size n=8000.