Implement individual preprocessing and modelling of each load-dataset
I.e.: The ability to preprocess and model the load-timeseries for each load-point (consumer) individually and storing them in a dictionary keyed by ID.
Also generalizes preprocessing to work on timeseries insted of structured data-dictionary-format, as well as move calculation of normal temperature and 3-day average to outside preprocessing-step, thereby requiring these as parts of the dict_data_ts-input. The functions calculating this may be moved for clarity.

Configs updated accordingly